Make union_outputs consume an iterator, not a slice
This avoids allocating extra Vecs in a couple of cases, and makes the
surrounding code simpler there.

The debug_assert I'm removing from union_outputs was redundant so long
as all the outputs had been used prior to this rewrite. If some were
unused then the other debug_assert in this function won't get a chance
to check everything, but it's not worth counting the iterator elements
just for this check that isn't supposed to ever happen.